chore: Convert CoverageSummary to class
